[REF] sale_pdf_quote_builder: code review
Overrid default_get for default documents that are not related to any
sale.order.template, therefore they are added only at creation.

The other documents are managed in the _onchange_sale_order_template_id:
  - manual cleanup of old documents that are not available anymore
    note: without telling the user
  - if there is a template set, then add its default documents
    note: done every time you change the template, but default docs can
    always be deselected

Additionally, tests now use the Form to tests this feature as it is
expected to be used in the ui only. (and its the only way to trigger
onchange ^^).