[FIX] product_margin: correct product margin report
On a sale order, when creating a down payment invoice and/or adding a global discount, the product margin report is incorrect.

- Create a sales order with a product with a list price of 1000.
- Add a global discount of 50% and confirm the SO.
- Create a down payment invoice on the SO for 50% and confirm.
- Create a regular invoice on the SO and confirm.
- In the product margin report, the down payment/discount products are added as credit, which makes the total turnover 2000 despite the balance sheet indicating that we sold for 500.

In `product.product` of the module `product_margin`, the SQL query took the absolute value of the balance of every account move line of display type product, including discounts and down payment lines. However, those should be debited on the product margin report.

After this commit, the discount and down payment products are debited on the product margin report so that it corresponds to the balance sheet.

**opw-4342691**