[IMP] website_sale: other solution
For the base case rule (ie. access to published categories), rule is
non-global and targets the three main groups:
- public
- portal
- user

This way it is "global" in the sense that any user should be part of at
least one of these three group.
Note: except for admins, but they have full access.

The second rule gives full access to website editors and admins.