[FIX] mail: properly truncate payload for push notifications
Steps to reproduce:

1. Activate desktop notifications.
2. Fastest way will be to send to an account A (with activated notis),
a message through discuss from an account B.
3. Additionally this affects the current flow too:
  3.1 When we try to notify a user following a heldesk team.
  3.2 We create an alias and set this to that current team.
  3.3 We simulate an email to that heldesk team, eg vip-support@test.com

Issue:

The _truncate_payload method does not work as expected because it does
not account for the post-encryption size increase in push_to_end_point.
The encryption process adds extra bytes to the payload header:

salt (16)
record_size (4)
sender_public_key.len (1)
sender_public_key (sender_public_key.len) usually around 65 bytes

This results in the payload exceeding the 4096-byte limit.
Solution:

To address this, we now reserve approximately 106 bytes for encryption
overhead and truncate the payload accordingly to ensure it remains
within the limit.

Additionally, the calculation of body_max_length was incorrect, as it
allowed for negative values, which sometimes resulted in an empty body.

The correct calculation is:
`3990 - payload_length``, which determines how many bytes need to be
removed to fit within the limit. Since this calculation can yield a
negative number, we adjust it by adding the `body_length`` to determine
the actual number of bytes that can remain in the body.

This approach ensures the payload remains within the 4096-byte limit
while preserving as much of the content as possible.

opw-3939019