[IMP] html_editor: add JS migration system
This commit adds the JS migration system that was previously implemented in
Knowledge to the standard. See
odoo/enterprise@0df669159aaa1a9631ab12a68758bf78954071f7

It was previously not necessary to use it in the standard because it only
affected Knowledge at the time. But from recent development it is now useful in
the standard as it now affects generic embedded components.

task-4489076