[FIX] portal: display template for portal invite
Steps to reproduce:
----

- Create a portal access to a user
- Go to the email template (Portal: User Invite)
- Preview

Issue:
---

The display was based on the user itself referencing to a partner. And
there was no name for the portal users.

Fix:
---

Changed the view so now the different values are based on the partner.Also
added the name to this function based on the partner name.

opw-4444729