[FIX] web_editor: avoid horizontal scrollbar in right panel
In edit mode, languages like Lithuanian or Spanish may cause a
horizontal scrollbar to appear in the right panel when toggling the grid
view. This issue arises because the Add Elements widget and the button
text are too large,leading to overflow.

This commit improves the CSS for we-buttons to ensure buttons adapt to
the available width. When multiple buttons are present, their widths are
now evenly distributed, preventing overflow and ensuring a consistent
layout across languages.

task-4378522