[FIX] mail: livechat from visitor stays open in mobile
Before this commit, when using livechat by visitor in mobile,
any page reload would remove access to the livechat.

Steps to reproduce:
- access to website with livechat installed on mobile
- click on livechat button (bubble icon in bottom right)
- send a message
- reload the page

=> the livechat is not open, not even as a chat bubble.

This prevents visitor from access the livechat again, except if
user finds a workaround to open the same page in non-mobile mode
(e.g. landscape mode or disabling high-dpi).

This happens because livechat relies on discuss chat hub, and
there's some code to prevent server-side synchronisation of chat
windows and bubbles in mobile. As a reminder, conversation in
mobile are shown as fullscreen chat windows.
This code was preventing the opening of chat window on livechat
for visitor, leading to prevention of continuing to live chat
on the device.

This commit fixes the issue by limiting the disabling of chat
window synchronisation to the web client. This is made possible
by relying on heuristics of presence of `im_livechat.livechat`
service: non-mobile mode or presence of livechat service enables
chat hub, instead fo non-mobile mode before. This allows livechat
visitor to keep livechat open.

Livechat in mobile takes the fullscreen, so with this fix the
livechat staying open would prevent navigation to the website unless
the livechat has ended. This is not great because most of the time
we want to livechat and still navigate on the website, even on mobile
device. This commit also enables folding of chat windows in mobile
for livechat visitors.

opw-4423556
opw-4488830