[IMP] website: add a new scrolling mode to the carousel snippet
Add a single card scroll mode. Previously, the user could only scroll
all items at once. Now, they can choose between two sliding modes:
- All: Default behavior as before (all items at once)
- Single: Scroll only one item at a time.

It was important not to modify the original carousel and to stick with
the Bootstrap carousel. Therefore, a new CSS class was introduced to
modify its behavior and enable a single-scroll mode.

This version works well for users with a good internet connection.
However, for users with slower connections, the animation may break
when the images finish loading. To address this issue, a preloading
mechanism was implemented.

Also this commit add a new option named Cycle, responsible for the
carousel cycle:
- Yes: The carousel will cycle (default, as before)
- No: The carousel won't let you cycle, stopping from sliding when
the last item is reached

task-4309599